-Split Int512 file.
